requirements.txt
accelerate==1.10.0
altair==5.5.0
annotated-types==0.6.0
anyio==4.7.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
arviz==0.21.0
asttokens==3.0.0
async-lru==2.0.5
attrs==24.3.0
babel==2.17.0
beautifulsoup4==4.13.5
bleach==6.2.0
blinker==1.9.0
Bottleneck==1.4.2
brotlicffi==1.0.9.2
cached-property==1.5.2
cachetools==5.5.1
catboost==1.2.8
category_encoders==2.8.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
cloudpickle==3.1.1
comm==0.2.3
cons==0.4.6
contourpy==1.3.2
cycler==0.12.1
dash==3.1.1
debugpy==1.8.11
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
etuples==0.3.9
exceptiongroup==1.3.0
executing==2.2.0
fastjsonschema==2.20.0
filelock==3.18.0
Flask==3.1.0
fonttools==4.59.0
fqdn==1.5.1
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.43
graphviz==0.20.1
greenlet==3.2.4
h11==0.16.0
h5netcdf==1.2.0
h5py==3.14.0
hf-xet==1.1.7
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.7
imbalanced-learn==0.13.0
importlib_metadata==8.7.0
ipykernel==6.30.0
ipython==8.37.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.6.1
joblib==1.4.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2023.7.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.8
lark==1.3.0
lightgbm==4.4.0
logical-unification==0.4.6
markdown-it-py==2.2.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.0
miniKanren==1.0.3
mistune==3.1.4
mpi4py==4.0.3
mpmath==1.3.0
multipledispatch==0.6.0
narwhals==1.31.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest_asyncio==1.6.0
networkx==3.4.2
notebook==7.4.5
notebook_shim==0.2.4
numexpr==2.11.0
numpy==2.2.5
openai==1.99.7
overrides==7.7.0
packaging==25.0
pandas==2.3.1
pandocfilters==1.5.0
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.3.0
pip==25.1
platformdirs==4.3.8
plotly==6.2.0
prometheus_client==0.23.1
prompt_toolkit==3.0.51
protobuf==4.25.3
psutil==5.9.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==16.1.0
pycparser==2.21
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pymc==5.25.1
pyparsing==3.2.3
PySocks==1.7.1
pytensor==2.31.7
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.30.2
regex==2025.7.34
requests==2.32.4
retrying==1.3.3
rfc3339_validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==13.9.4
rpds-py==0.22.3
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==72.1.0
six==1.17.0
sklearn-compat==0.1.3
smmap==5.0.2
sniffio==1.3.0
soupsieve==2.8
SQLAlchemy==2.0.43
stack_data==0.6.3
statsmodels==0.14.5
streamlit==1.49.1
sympy==1.14.0
tenacity==9.0.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.4.0
tokenizers==0.21.4
toml==0.10.2
tomli==2.2.1
toolz==1.0.0
torch==2.7.1
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.55.2
types-python-dateutil==2.9.0.20250822
typing_extensions==4.14.1
typing-inspection==0.4.0
typing_utils==0.1.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
watchdog==4.0.2
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
xarray==2025.4.0
xarray-einstats==0.8.0
xgboost==2.1.2
zipp==3.23.0
